refactor: 💡 move InputConnection extension methods to utils
